Move Tests from "shouldRun" to "discriminators" - move discriminator tests from "shouldRun" to "discriminators"
